renamed migrations

+++ database/migrations/2019_016_13_205000_create_pivot_team_user_table.php
@@ -23,8 +23,8 @@ class CreatePivotTeamUserTable extends Migration
 
 			Schema::create('team_user', function (Blueprint $table) {
 				$table->increments('id');
-				$table->unsignedInteger('team_id');
-				$table->unsignedInteger('user_id');
+				$table->unsignedBigInteger('team_id');
+				$table->unsignedBigInteger('user_id');
 
 				$table->foreign('team_id')
 						->references('id')
